set attachments subcommand
